requirements.txt
# Requirements for the Electra FastAPI backend
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
python-dotenv>=1.1.1
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
VeraGridEngine>=5.4.4
